Updating #105 (Tests) with develop

+++ src/main/java/ISPP/G5/INDVELOPERS/cloud/CloudStorageController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,13 +14,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+@CrossOrigin("*")
 @RestController
 @RequestMapping("/file")
 public class CloudStorageController {
 
     @Autowired
     private CloudStorageService cloudStorageService;
-
+    
     @PostMapping("/upload")
     public ResponseEntity<String> uploadFile(@RequestParam(value = "file") MultipartFile file) {
         return new ResponseEntity<>(cloudStorageService.uploadFile(file), HttpStatus.OK);
@@ -29,16 +31,31 @@ public class CloudStorageController {
     public ResponseEntity<ByteArrayResource> downloadFile(@PathVariable String fileName) {
         byte[] data = cloudStorageService.downloadFile(fileName);
         ByteArrayResource resource = new ByteArrayResource(data);
-        return ResponseEntity
-                .ok()
-                .contentLength(data.length)
-                .header("Content-type", "application/octet-stream")
-                .header("Content-disposition", "attachment; filename=\"" + fileName + "\"")
-                .body(resource);
+
+        if(this.cloudStorageService.canDownloadGame(fileName)) {
+            return ResponseEntity
+                    .ok()
+                    .contentLength(data.length)
+                    .header("Content-type", "application/octet-stream")
+                    .header("Content-disposition", "attachment; filename=\"" + fileName + "\"")
+                    .body(resource);
+        }else {
+            return ResponseEntity
+            		.status(HttpStatus.FORBIDDEN)
+                    .contentLength(data.length)
+                    .header("Content-type", "application/octet-stream")
+                    .header("Content-disposition", "attachment; filename=\"" + fileName + "\"")
+                    .body(null);
+        }
     }
 
     @DeleteMapping("/delete/{fileName}")
     public ResponseEntity<String> deleteFile(@PathVariable String fileName) {
-        return new ResponseEntity<>(cloudStorageService.deleteFile(fileName), HttpStatus.OK);
+        if(this.cloudStorageService.isOwnerOfTheFile(fileName)) {
+            return new ResponseEntity<>(cloudStorageService.deleteFile(fileName), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("You are not the owner of the file", HttpStatus.FORBIDDEN);
+
+        }
     }
 }
